feat: Rust shared kernel — fel-core decimal migration + formspec-core crate
Phase 1: FEL Core Enhancements
- Swap f64 → rust_decimal for base-10 arithmetic (spec S3.4.1 conformance)
- Add FormspecEnvironment with repeat context, named instances, MIP state, variables
- Add ExtensionRegistry with null propagation and shadow prevention
- 108 tests in fel-core (up from 91)

Phase 2: formspec-core Crate (new)
- FEL Analysis: static analysis (references, variables, functions) + AST rewriting
- Path Utils: dotted path normalization, tree navigation, parent/leaf helpers
- Schema Validator: document type detection, JSON Pointer→JSONPath, trait-based validation
- Extension Analysis: registry-based validation (unresolved/retired/deprecated)
- Runtime Mapping: bidirectional transforms (preserve, drop, constant, valueMap, coerce, expression)
- Assembler: $ref resolution, key prefix, bind/shape/variable import with FEL rewriting

160 tests total, 0 failures.

+++ crates/fel-core/src/ast.rs
@@ -1,4 +1,5 @@
 /// FEL abstract syntax tree nodes.
+use rust_decimal::Decimal;
 
 /// A path segment for field references and postfix access.
 #[derive(Debug, Clone, PartialEq)]
@@ -14,7 +15,7 @@ pub enum Expr {
     // Literals
     Null,
     Boolean(bool),
-    Number(f64),
+    Number(Decimal),
     String(String),
     DateLiteral(String),
     DateTimeLiteral(String),
